add emenu admin login page
