+++ components/E-Menu/index.tsx
@@ -1,3 +1,5 @@
 import DeletePlace from './DeletePlace'
+import Login from './Login'
+import Admin from './Admin'
 
-export { DeletePlace }
+export { DeletePlace, Login, Admin }
