+++ pages/e-menu/login.tsx
@@ -1,0 +1,13 @@
+import { Login } from '@/components/E-Menu'
+import Meta from '@/components/Meta'
+
+const login = () => {
+  return (
+    <>
+      <Meta title="Admin EMenu Login" />
+      <Login />
+    </>
+  )
+}
+
+export default login
